Port LiquidiyLayer PRs to v2 (#1497)
* Add PortalAdapter to LiquidityLayer (#1290)

* Add PortalAdapter to LiquidityLayer

* newlines

* Fix up circle relayer

* Dont require context when getting args

* Fix up context argument parsing

* Liquidity Layer Relayer Infras (#1292)

* PR review

* PR review

* Use redeployed CCTP contracts (#1482)

* Use redeployed CCTP contracts

* Fix build

* Rename more things to TokenMessenger and fix nonceHash calculation

* New artifacts

* Fix build

* Deploy relaers

* Gas snapshot

* PR review

* Add new line

+++ solidity/contracts/mock/MockCircleMessageTransmitter.sol
@@ -20,7 +20,7 @@ contract MockCircleMessageTransmitter is ICircleMessageTransmitter {
         success = true;
     }
 
-    function hashSourceAndNonce(uint32 _source, uint256 _nonce)
+    function hashSourceAndNonce(uint32 _source, uint64 _nonce)
         public
         pure
         returns (bytes32)
